Only update box if a hash if key/values is given
Prior to this commit, the update command would send an empty hash to
Vagrant Cloud if the Box.update method was called with no data. This
commit fixes that by returning early if an empty hash was given. It also
fails over to using the class instance variables for the box org and
name, if not given.