Pass token and custom server information
The primary purpose of this change is to pass the authentication information through during object creation in the deprecated helpers.  Upon furthur review, if a custom server is defined in the account, it would also suffer from the same issue.  A new accessor is defined in the client to expose the url_base attribute.