Add support for HCP authentication
Enables API requests to use a static access token value or to
generate tokens using an HCP service principal if the expected
environment variables are set. The authentication header is
no longer set at client initialization and is instead added to
the connection just prior to the request. This will allow the
token to be properly regenerated during long running processes
where the token may have expired.